fix(dd): armhf -Wconversion — ibs/obs use size_t
ibs/obs were uint64_t, but they feed std::vector ctors, read()/memset()
sizes and got==ibs comparisons — all size_t. On 32-bit armhf (size_t =
unsigned int) the uint64_t->size_t narrowing tripped -Werror=conversion
in CI. ibs/obs are now size_t (a block size never realistically exceeds
SIZE_MAX); count/skip/seek stay uint64_t (may be large, used for loops
and lseek off_t). Parse assignments cast n->size_t.

Verified locally: armhf cross-compile green, native ctest 489 + dd 10 green.
This was the only CI failure on feat/phase4-sweep (native/release-size passed).

+++ src/applets/dd.cpp
@@ -66,8 +66,8 @@ auto to_lower(std::string_view s) -> std::string {
 }
 
 struct DdOpts {
-    std::uint64_t ibs = 512;
-    std::uint64_t obs = 512;
+    std::size_t ibs = 512;
+    std::size_t obs = 512;
     std::uint64_t count = 0;        // 0 = unlimited
     std::uint64_t skip = 0;
     std::uint64_t seek = 0;
@@ -137,13 +137,13 @@ auto dd_main(int argc, char* argv[]) -> int {
 
         if (key == "bs") {
             if (!parse_num(val, n) || n == 0) { CFBOX_ERR("dd", "invalid bs: %s", std::string(val).c_str()); return 1; }
-            o.ibs = o.obs = n;
+            o.ibs = o.obs = static_cast<std::size_t>(n);
         } else if (key == "ibs") {
             if (!parse_num(val, n) || n == 0) { CFBOX_ERR("dd", "invalid ibs: %s", std::string(val).c_str()); return 1; }
-            o.ibs = n;
+            o.ibs = static_cast<std::size_t>(n);
         } else if (key == "obs") {
             if (!parse_num(val, n) || n == 0) { CFBOX_ERR("dd", "invalid obs: %s", std::string(val).c_str()); return 1; }
-            o.obs = n;
+            o.obs = static_cast<std::size_t>(n);
         } else if (key == "count" || key == "skip" || key == "seek") {
             if (!parse_num(val, n)) { CFBOX_ERR("dd", "invalid %.*s: %s", static_cast<int>(key.size()), key.data(), std::string(val).c_str()); return 1; }
             if (key == "count") o.count = n;
